Add a (hard-coded) way to filter and organize ping types
This is just a framework that allows the server to tell the client which ping types it wants to actually allow on the server, and in which order they should be in the ping selector wheel. The filter acts as a whitelist system so that if they do not want the default "vanilla" ping types provided by the mod, they can omit them from the (soon to be) server configuration file.

The goal would be to also have a client-sided override over the order of the ping types in case they would prefer them sorted another way, but the server could provide the default sort order. Regardless though, the client should never be allowed to send a ping that is not in the server-defined filter, even if they're in the registry.